pulled apart declarations and implementations for those objects that I need in FalconIC. Fixed some memory leaks, not much else.
